fixed bug that would crash the drive controller at the end of a path

+++ src/controls/drive_controller/include/drive_controller/drive_controller.h
@@ -41,7 +41,7 @@ public:
    void haltAndAbort();
    bool update(DriveController_ns::robot_state_vector sv, double dt);
    double getPClosestT();
-   double getPPaths();
+   int getPPaths();
 //protected:
    double angleDiff(double angle1,double angle2);
    std::pair<double, double> speedSteeringControl(double speed, double steering,
